// src/app/shared/models/bp.model.ts
export class Bp {
  bp_nombre: string;
  bp_perfilamiento: string;
  bp_tipo: string;
  bp_status: string;
  bp_fecha_inicio: any;
  bp_fecha_fin: any;
  bp_presentacion: string;
  bp_firma_contrato_esa: string;
  bp_kick_off: string;
  bp_plan_negocios: string;
  bp_comercial: string;
  bp_readiness: string;
  bp_com_prensa: string;
  bp_quarter_closed: string;
  bp_dia_cadencia: string;
  bp_sig_pasos: string;
  bp_cad_sem: string;
  bp_progre: string;
  bp_cert: any;
  bp_opp: any;
}